[backend] 1. 그룹 수정  - 특정 주차 컴프레셔 리스트를 빈 배열로 전달 시 수정 안되는 이슈 수정  - min, max 소수점 적용

+++ ems-ing/src/main/java/com/markcha/ems/dto/schedule/ScheduleDto.java
@@ -41,8 +41,8 @@ public class ScheduleDto {
     private String type;
     private LocalTime startTime;
     private LocalTime stopTime;
-    private Integer min;
-    private Integer max;
+    private Double min;
+    private Double max;
     @JsonIgnore
     private LocalDateTime updated;
     private List<DayOfWeekDto> dayOfWeeks;
